Added function to debug futures thread and inserted into database
Added function to debug the errors occuring in the futures threads.
Without that, the threads only mention the type of error, and not the
error itself. This commit also allows for insertion of data into the
mysql table which was not possible earlier due to concurrent calls to
the database. This problem has been fixed with the help of Connection
Pooling of MySQL connector